Snapcam: preview cover hide issue
Issue: whe switch module from photo to video, the preview cover will
hide when capture result arrived, but it may be from last module.
So cover hidden at a wrong time.
Fix: add the module info to the preview request tag, and then when need
to hide cover, we check the module info in the tag, whether it is from
last module.

Change-Id: I6af165799c4f46795b73a8f84fba289364e78bed