SnapdragonCamera: CaptureModule: send the EOS in advance
in this project, mpeg4 node and the video record sound play
consume long duration. it caused the EOS cannot be transfer
to camera frameowork on time and then cannot be transfer
to camera hal. finally this cause the camera EIS node buffer
cannot be return back to framework and then stop video recording
failed. this change used to send EOS before play video sound
to save latency of EOS.

Change-Id: I894a7ee9b7a1dd11a1faf2ecd74483c1a5b20691